Mock Details WIP

+++ client/mockpit-ui/src/app/services/mock.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { ConfigService } from './config.service';
-import { MockResponse } from '../models/mock/mock.model';
+import { Mock, MockResponse } from '../models/mock/mock.model';
 
 @Injectable({
   providedIn: 'root'
@@ -22,8 +22,16 @@ export class MockService {
     return this.http.get(this.backendUrl + "/native/api/mocks/" +id);
   }
 
-  public deleteMockById(id: number){
-    return this.http.delete(this.backendUrl + "/native/api/mocks/"+id);
+  public saveMock(mock:Mock) : Observable<MockResponse>{
+    return this.http.post(this.backendUrl + "/native/api/mocks", mock) as Observable<MockResponse>;
+  }
+
+  public deleteMockById(id?: number){
+    if(id){
+      return this.http.delete(this.backendUrl + "/native/api/mocks/"+id);
+    } else {
+      return of("Mock id not provided");
+    }
   }
 
   public deleteAllMocks(){
